Method for table_detect()
Method for table_detect() function, uses beautifulsoup string arguements in combination w/ regex(need to figure out how to write regex patterns for complex variable situations like splitting). For this function to work , need to be able to find certain strings based on html tags in header rows , if that doesn't work search the row cells for splitting/ carbon type patterns. If that doesn't work, thought of way for converting numbers in columns to float, making list with right numbers(just need most for average) and checking range for values; if 1-13 HNMR, if 14-100 CNMR. If all fails, then ask user for table type.