Trying to get different tables to work; Figuring out how to get final data into output JSON
Testing and modifying the code to sucessfully parse other table types that are quite different. Have many different searches and test, but seems to fail when certain info is not found. Can easily make this program work for this one table, just trying to make it capable of doing many types that unfortunately differ enough to cause such specific clunky code to parse data. Everytime I try a new table it shows many flaws in program. Even if table type is the right one, the creator can add in one puncuation or extra data that breaks code.