Fixing encoding errors
Beatifulsoup will convert the html table encoded in UTF-8 on the webpage(UTF-8 is the standard) to unicode and output soup as UTF-8 because
"BeautifulSoup internally uses the sub-library called Unicode, Dammit to detect a document’s encoding and then convert it into Unicode"

"The output from a BeautifulSoup is UTF-8 document, irrespective of the entered document to BeautifulSoup"

This means that some characters that can't convert from utf-8 to unicode are lost even when converted back to utf-8. Have to find  out how Beautifulsoup converts unknowns to unicode, so I might be able to encode back properly into utf-8 and bypass removing each encoding error possible. Is there another html parse that doesnt convert internally to unicode?

Also an error where HTML entities are converted to Unicode characters. &nbsp; inside of string within html <tag> become part of the string. Will likely have to just search and remove if not alone; converted to unicode.