Testing
All the failures caused errors in determining which variables are present at the end before adding to csv(Seems to be happening for Carbon only or Both w/o C type(HNMR only works fine)).  Need to reconfigure how to add data to csv so that blank list are not added ruining the formatting.  Seems like currently unable to successfully parse each of the three types: H/C or Both. Want to have it so that if float_hspec/float_cspec are empty can try to use float_average to determine if C/H present and create float_hspec/float_cspec. Then for each case(If certain variable are missing) can then add to csv. Initially tried this, but run into issue, must be a better way. ** May need to use table detect for this purpose, but not sure if function will be clunky and not work. But should try having it return if H or C or Both, then based on return of True/False pick which method to use to append to csv**
Also, fix so that comp_num can try looking at headers, then if numbers in headers to compare, since headers may only be 1 while num_headers = 4
, currently will return 1 even though 5 comps(due to swapped headers).