Testing
Almost verything is working and fixed the small reasons for errors. But cannot seem to fix CNMR only case, as no hspec, but blank list appears causing program to think hspec is present; only effected part is the headers. Only reason why failures for test. The parser is not adding CNMR only to the csv file and with no results, nothing to compare to the expected.