Test for tables with other data included
Will parse for information that program can detect, otherwise left out and will require editing the code to parse as well. Issues mainly arrise when the format of the table differs in the structuring of the format not so much random info in the cells or other columns(ex. rowspan = 2).