Updated regex.
Adding more tests. 6/56 fails - Same as before.

Determined tests.test_souping.test_parse.html parsing error was due to eq being in the cell. For some reason if eq in the cell, it inserts a new row into the table beneth the eq cell as if it were a multi-data cell.